requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
flask==3.1.0
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
kafka-python==2.0.2
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-mistralai==0.2.4
langchain-openai==0.2.14
langchain-text-splitters==0.3.5
langsmith==0.2.10
MarkupSafe==3.0.2
marshmallow==3.24.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.59.4
orjson==3.10.13
packaging==24.2
propcache==0.2.1
pydantic==2.10.4
pydantic-core==2.27.2
pydantic-settings==2.7.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.67.1
typing-extensions==4.12.2
typing-inspect==0.9.0
urllib3==2.3.0
werkzeug==3.1.3
yarl==1.18.3
zipp==3.21.0
